Refactor test names and add minor fixes
