+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/field/type/MultiLevelFieldAccessInFunction2.bal
