+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/field/type/MultiLevelFieldInDifferentFileAccessInFunction1.bal
