+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/field/type/MultiLevelFieldInDifferentFileAccessInService2.bal
