+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/field/type/MultiLevelFieldInDifferentPackageAccessInFunction2.bal
