+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/field/type/SingleLevelFieldAccessInFunction2.bal
