+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/field/type/SingleLevelFieldAccessInService1.bal
