+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/field/type/SingleLevelFieldAccessInService2.bal
