+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/field/type/SingleLevelFieldInDifferentPackageAccessInFunction1.bal
