+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/field/type/SingleLevelFieldInDifferentPackageAccessInService2.bal
