+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/function/FunctionInDifferentPackageInvokeInFunction.bal
@@ -1,5 +1,5 @@
-import org/test;
+import test;
 
-function test () {
+function test() {
     test:/*ref*/fun();
 }
