+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/type/RecordInDifferentFileInFunction.bal
