+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/type/RecordInDifferentFileInService.bal
