+++ tool-plugins/intellij/src/test/resources/testData/psi/resolve/type/RecordInDifferentPackageInService.bal
